requirements.txt
fastapi
uvicorn[standard]
pdfplumber
anthropic
sentence-transformers
faiss-cpu
numpy
python-multipart
python-dotenv
